requirements.txt
rawpy
Pillow
openai